feat(StringUtil) : String 을 int 로 변환
- 테스트 코드 작성
